Fixed the issue about non-integer indexfor numpy array

+++ hmtk/plotting/seismicity/max_magnitude/cumulative_moment.py
@@ -64,7 +64,7 @@ def plot_cumulative_moment(year, mag):
     # Calculate seismic moment
     m_o = 10. ** (9.05 + 1.5 * mag)
     year_range = np.arange(np.min(year), np.max(year) + 1, 1)
-    nyr = np.float(np.shape(year_range)[0])
+    nyr = np.int(np.shape(year_range)[0])
     morate = np.zeros(nyr, dtype=float)
     # Get moment release per year
     for loc, tyr in enumerate(year_range):
@@ -72,7 +72,7 @@ def plot_cumulative_moment(year, mag):
         if np.sum(idx) > 0:
             # Some moment release in that year
             morate[loc] = np.sum(m_o[idx])
-    ave_morate = np.sum(morate) / nyr
+    ave_morate = np.sum(morate) / float(nyr)
 
     # Average moment rate vector
     exp_morate = np.cumsum(ave_morate * np.ones(nyr))
